feat: 积分现金双报、青龙联动、Bark推送与列表体验升级
支持可选 current_cash 上报与可兑换库存高亮；同步青龙启停/cron，
增加 Bark 未报提醒、设置分栏与数据库备份，并修复移动端导航切换问题。

+++ app/routers/stock.py
@@ -205,7 +205,7 @@ def get_stock_center(db: Session = Depends(get_db)):
             models.PointsHistory.report_time == latest_points_subquery.c.max_time,
         ),
     ).group_by(models.PointsHistory.program_id).all()
-    max_points_map = {row.program_id: int(row.max_user_points or 0) for row in max_points_rows}
+    max_points_map = {row.program_id: float(row.max_user_points or 0) for row in max_points_rows}
 
     program_rows = db.query(models.MiniProgram).order_by(
         desc(models.MiniProgram.sort_order),
